Stop using smallint as primary key type

+++ db/sql/007_create_grants.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS tbl_fact_grants(
         id int UNSIGNED NOT NULL auto_increment PRIMARY KEY,
         grant_number varchar(20),
         title varchar(30),
-        source_id smallint UNSIGNED,
+        source_id int UNSIGNED,
         original_amt float(7,2),
         balance float(7,2),
         status ENUM('PENDING', 'APPROVED', 'DENIED'),
